Implement databinding for currency item layout
