+++ app/src/main/res/values/dimens.xml
@@ -7,4 +7,5 @@
 
     <dimen name="text_body">16sp</dimen>
     <dimen name="text_body_small">14sp</dimen>
+    <dimen name="title_header_small">20sp</dimen>
 </resources>
